Current Code is broken
The issue is that within table it is still trying to parse through the SessionMemory. The table must call each calculation on it's own and to do that it must be able to pass both metalLossList and the InputList to each calculation. This has already been done for ILIB31GModifiedCalculation